refactored human model
