+++ opendemic/webhook/telegram/controller.py
@@ -7,7 +7,7 @@ from telebot.types import Update
 from flask import Blueprint, request, abort
 import datetime
 from opendemic.webhook.telegram.model import log_action, log_sent_message
-from opendemic.human.human import Human
+from opendemic.human.model import Human
 from opendemic.webhook.telegram.api_helpers import get_telegram_menu, get_telegram_bot_instance, \
     get_webhook_update, make_reply_keyboard_markup
 from opendemic.webhook.telegram.model import TelegramCommand
